Indexing is nearly fully operational.
Existing database entries are just updated and only new ones added

Every database entry has a mtime date of the file and a timestamp when the
entry was last updated. To find entries that does no longer exists in the
filesystem, one has just to find entries with an outdated dbmtime field.

But it fails to create a database entry when the tags aren't readable. It is
marked with a fixed me.