Implemented an abstract TrackSource interface
Index is now independent of the source. By using two interfaces defined in
source.go TrackInfo and TrackSource it isn't important anymore, where the tracks
come frome.

Also the crawling enginge (source) is now in charge of gettings the tags on
demand TrackInfo.Tags().

Needs to be done:
The filecrawler closes the channel, when it's done. This is stupid, if other
source want to plug in. Fix that!